Move config out of cli package, rename from core to config. Move AppCommands to commands.go
