+++ config/config.go
@@ -1,4 +1,4 @@
-package core
+package config
 
 import (
 	"os"
